gateway/requirements.txt
alembic==0.9.9
cryptography==2.3
Flask==1.0.2
#flask-apispec==0.7.0
Flask-Cors==3.0.4
#Flask-HTTPAuth==3.2.3
#flask-marshmallow==0.9.0
flask-nameko==1.4.0
#Flask-RESTful==0.3.6
#flask-restful-swagger-2==0.35
Flask-Script==2.0.6
Flask-SQLAlchemy
gunicorn==19.9.0
itsdangerous==0.24
#marshmallow==3.0.0b8
passlib
PyJWT==1.6.4
psycopg2
marshmallow>3
